Fix warnings associated with deprecated uses of iOS frameworks.  Also warnings with 32-bit/64-bit compiler.

+++ LBYouTubeView/LBYouTubeExtractor.m
@@ -196,7 +196,7 @@ NSInteger const LBYouTubePlayerExtractorErrorCodeNoJSONData   =    3;
                     streamURL = [[videos objectAtIndex:0] objectForKey:streamURLKey];
                 }
                 else if (self.quality == LBYouTubeVideoQualityMedium) {
-                    unsigned int index = MAX(0, videos.count-2);
+                    NSUInteger index = MAX(0, videos.count-2);
                     streamURL = [[videos objectAtIndex:index] objectForKey:streamURLKey];
                 }
                 else {
